+++ Example/SitumWayfinding/WayfindingController.swift
@@ -12,7 +12,7 @@ import SitumWayfinding
 import SitumSDK
 import GoogleMaps
 
-class WayfindingController: UIViewController {
+class WayfindingController: UIViewController, OnPoiSelectionListener, OnFloorChangeListener {
     
     @IBOutlet var containerView: UIView!
     
@@ -39,6 +39,8 @@ class WayfindingController: UIViewController {
             options.buildingDetector = .SITBLE
             locationRequest.outdoorLocationOptions = options
         }
+        self.library?.setOnPoiSelectionListener(listener: self)
+        self.library?.setOnFloorChangeListener(listener:self)
             
         do {
             try self.library!.load()
@@ -52,6 +54,18 @@ class WayfindingController: UIViewController {
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
+    }
+    
+    // MARK: Wayfinding Delegate
+    func onPoiDeselected(building: SITBuilding) {
+        print("onPoiDeselected app")
+    }
+    
+    func onPoiSelected(poi: SITPOI, level: SITFloor, building: SITBuilding) {
+        print("onPoiSelected")
+    }
+    func onFloorChanged(from: SITFloor, to: SITFloor, building: SITBuilding) {
+        print("onFloorChanged from \(from.floor) to \(to.floor)")
     }
 }